+++ src/Frontend/FluentCMS.Web.UI.Components/Components/Form/FormInput/FormInput.razor.cs
@@ -5,7 +5,7 @@ namespace FluentCMS.Web.UI.Components;
 public partial class FormInput
 {
     [Parameter]
-    public int Cols { get; set; } = 12;
+    public decimal Cols { get; set; } = 12;
 
     protected override bool TryParseValueFromString(string? value, out string? result, [NotNullWhen(false)] out string? validationErrorMessage)
     {
